changes to renaming errors
