+++ assignment2/src/components/Dashboard.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { getTopGainersLosers } from '../store/reducers/stocks.js'; 
-import { gainersSelector, losersSelector } from '../store/selectors/StocksSelectors.js';
+import { gainersSelector, losersSelector, loadingSelector, errorSelector } from '../store/selectors/StocksSelectors.js';
 import GainersTable from './GainersTable';
 import LosersTable from './LosersTable';
 import data from '../data/demo.json'; // for my own demo data
@@ -11,7 +11,8 @@ const Dashboard = () => {
   // const { gainers, losers, loading, error } = useSelector((state) => state.stocks);
   const gainers = useSelector(gainersSelector);
   const losers = useSelector(losersSelector);
-
+  const loading = useSelector(loadingSelector);
+  const error = useSelector(errorSelector);
 
   useEffect(() => {
     dispatch(getTopGainersLosers());
@@ -24,11 +25,11 @@ const Dashboard = () => {
     <div className='flex justify-between items-center '>
      <div className='flex flex-col justify-center items-center mr-10'>
         <h2 className='text-3xl font-semi-bold text-lime-600 mb-7'>Top Gainers</h2>
-        <GainersTable gainers={data.top_gainers} />   {/* gainers = {gainers}*/}
+        <GainersTable gainers={gainers} />   {/* gainers = {gainers}*/}
      </div>
      <div className='flex flex-col justify-center items-center ml-10'>
         <h2 className='text-3xl font-semi-bold text-red-700 mb-7'>Top Losers</h2>
-        <LosersTable losers={data.top_losers} />  {/* losers = {losers}*/}
+        <LosersTable losers={losers} />  {/* losers = {data.top_losers}*/}
      </div> 
     </div>
   );
