+++ assignment2/src/main.jsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import {Provider} from "react-redux"
-import {store} from "./redux/store.js"
+import {store} from "./store/store.js"
 import App from './App.jsx'
 
 createRoot(document.getElementById('root')).render(
